+++ sketch-compose-core/src/commonMain/kotlin/com/github/panpf/sketch/transition/TransitionComposeTarget.kt
@@ -18,6 +18,10 @@
 package com.github.panpf.sketch.transition
 
 import androidx.compose.ui.graphics.painter.Painter
+import androidx.compose.ui.layout.ContentScale
+import com.github.panpf.sketch.resize.ScaleDecider
+import com.github.panpf.sketch.util.fitScale
+import com.github.panpf.sketch.util.toScale
 
 interface TransitionComposeTarget : TransitionTarget {
 
@@ -25,4 +29,16 @@ interface TransitionComposeTarget : TransitionTarget {
      * The component's current [Painter].
      */
     val painter: Painter?
+
+    /**
+     * The [ContentScale] of the component that this target is applied to.
+     */
+    val contentScale: ContentScale
+
+    @Deprecated("Please use contentScale instead and will be deleted in the future")
+    override val fitScale: Boolean
+        get() = contentScale.fitScale
+
+    // TODO Combined with alignment to convert more precisely Scale
+    override fun getScaleDecider(): ScaleDecider? = ScaleDecider(contentScale.toScale())
 }