+++ sketch-view-core/src/main/kotlin/com/github/panpf/sketch/target/TransitionViewTarget.kt
@@ -18,8 +18,13 @@
 package com.github.panpf.sketch.target
 
 import android.graphics.drawable.Drawable
+import android.widget.ImageView
+import android.widget.ImageView.ScaleType
+import com.github.panpf.sketch.resize.ScaleDecider
 import com.github.panpf.sketch.transition.Transition
 import com.github.panpf.sketch.transition.TransitionTarget
+import com.github.panpf.sketch.util.fitScale
+import com.github.panpf.sketch.util.toScale
 
 /**
  * A [Target] that supports applying [Transition]s.
@@ -30,4 +35,15 @@ interface TransitionViewTarget : TransitionTarget {
      * The current [Drawable].
      */
     val drawable: Drawable?
+
+    /**
+     * The [ScaleType] of the [ImageView] that this target is applied to.
+     */
+    val scaleType: ScaleType
+
+    @Deprecated("Please use scaleType instead and will be deleted in the future")
+    override val fitScale: Boolean
+        get() = scaleType.fitScale
+
+    override fun getScaleDecider(): ScaleDecider? = ScaleDecider(scaleType.toScale())
 }